+++ src/main/java/com/gitlab/cdagaming/craftpresence/config/gui/MainGui.java
@@ -350,7 +350,7 @@ public class MainGui extends ExtendedScreen {
 
         // noinspection RedundantSuppression,ConstantConditions,MismatchedStringCase
         if (!ModUtils.VERSION_TYPE.equalsIgnoreCase("release")) {
-            drawString(mc.fontRenderer, releaseNotice, (width / 2) - 30 - (StringUtils.getStringWidth(releaseNotice) / 2), height - 85, 0xFFFFFF);
+            drawString(mc.fontRenderer, releaseNotice, (width / 2) - (StringUtils.getStringWidth(releaseNotice) / 2), height - 85, 0xFFFFFF);
         }
 
         syncRenderStates();